Refactor secrets into a domain object.
This moves secrets into a concrete type that allows validating against request bodies and provided signatures. This allows a lot of flexibility when using this package as a handler for multiple secrets. By making the handler use a Resolver for secrets, it allows developers to consume webhooks for multiple projects.